% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_tune.R
\name{allo}
\alias{allo}
\title{C input calculation}
\usage{
allo(yield_MC, yield_CC, C_manure, HI, SB, RE, RB, Ccont = 0.43)
}
\value{
The arguments are:
}
\description{
Simple allometrics function to estimates of C input to the soil from information of yield
amounts. Calculations of total C (Mg/ha) deposited in top and sub soil
}
