% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basic_example}
\alias{basic_example}
\title{Basic rCTOOL example dataset}
\format{
A data frame.
}
\source{
Internal example dataset distributed with rCTOOL.
}
\usage{
basic_example
}
\description{
Example dataset used to demonstrate a basic rCTOOL simulation.
}
\keyword{datasets}
