% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physical_boundaries.R
\name{check_balance}
\alias{check_balance}
\title{check_balance}
\usage{
check_balance(ctool_output, cin_config, s_config)
}
\arguments{
\item{ctool_output}{Output from turnover model.}

\item{cin_config}{Carbon input configuration.}

\item{s_config}{Soil configuration.}
}
\value{
The input `ctool_output` data.frame, returned unchanged after
checking carbon balance consistency.
}
\description{
Check carbon balance consistency of a  simulation.
}
\details{
Computes the balance between initial soil carbon, cumulative carbon inputs,
final carbon stocks and cumulative CO2 emissions. The balance should be
zero or very close to zero.
}
