% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature.R
\name{convert_day_to_monthly_temperatures}
\alias{convert_day_to_monthly_temperatures}
\title{convert_day_to_monthly_temperatures}
\usage{
convert_day_to_monthly_temperatures(file)
}
\arguments{
\item{file}{either filepath or a dataframe containing the temperature data}
}
\value{
data.frame containing monthly temperatures as well as temperature range for a given daily temperature data
}
\description{
reads file, identify whether file is a filepath or a dataframe containing the daily temperature data
calculates monthly mean Tmax, Tmin, Tavg as well as temperature range
}
\examples{
convert_day_to_monthly_temperatures(file='./data/tempfile.csv')
convert_day_to_monthly_temperatures(file=temperature_df)
}
