% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates.R
\name{decay}
\alias{decay}
\title{Decay of carbon function'}
\usage{
decay(C0_t, k, tempCoefficient)
}
\value{
The arguments are:
}
\description{
The decay of carbon in each pool is described by first-order reaction kinetics
where we have an specific decay rate coefficient in each pool that affects the
actual C content in it (MgC/ha).The turnover is simulated in monthly time steps,
where th decay rate is modify by the temp_coef() function.
}
