% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{define_Cinputs}
\alias{define_Cinputs}
\title{define_Cinputs}
\usage{
define_Cinputs(
  management_filepath = NULL,
  Cin_top = NULL,
  Cin_sub = NULL,
  Cin_man = NULL,
  time_config = NULL
)
}
\arguments{
\item{management_filepath}{Either a filepath to a management template or
a data.frame containing `Cin_top`, `Cin_sub` and `Cin_man`.}

\item{Cin_top}{Annual carbon input from plant residues in the topsoil.}

\item{Cin_sub}{Annual carbon input from plant residues in the subsoil.}

\item{Cin_man}{Annual carbon input from manure.}

\item{time_config}{Time configuration object returned by `define_timeperiod()`}
}
\value{
A list containing `Cin_top`, `Cin_sub` and `Cin_man`.
}
\description{
Prepare annual carbon input (from plants and/or manure) configuration.
}
\examples{
time_config <- define_timeperiod(yr_start = 2006, yr_end = 2008)
define_Cinputs(
  Cin_top = c(2, 2, 2),
  Cin_sub = c(0.5, 0.5, 0.5),
  Cin_man = c(1, 1, 1),
  time_config = time_config
)
}
