% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physical_boundaries.R
\name{define_physical_boundaries}
\alias{define_physical_boundaries}
\title{define_physical_boundaries}
\usage{
define_physical_boundaries(value, min_limit, max_limit)
}
\arguments{
\item{value}{parameter value}

\item{min_limit}{Minimum allowed value.}

\item{max_limit}{Maximum allowed value.}
}
\value{
The input value if it lies within the specified physical bounds.
}
\description{
Check whether a parameter value lies within specified physical bounds.
}
\examples{
define_physical_boundaries(0.3, 0, 1)
}
