% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{define_timeperiod}
\alias{define_timeperiod}
\title{define_timeperiod}
\usage{
define_timeperiod(yr_start, yr_end)
}
\arguments{
\item{yr_start}{Initial simulation year.}

\item{yr_end}{Final nd simulation year.}
}
\value{
A list with two elements: `timeperiod`, a data.frame containing
monthly timesteps, years and annual timestep identifiers; and `steps`,
the total number of simulated months.
}
\description{
Create the monthly simulation time structure.
}
\examples{
 define_timeperiod(yr_start=2006, yr_end=2010)
}
