% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_soil_pools.R
\name{.clean_monthly_allocations}
\alias{.clean_monthly_allocations}
\title{.clean_monthly_allocations}
\usage{
.clean_monthly_allocations(m_config)
}
\arguments{
\item{m_config}{Management configuration list.}
}
\value{
A management configuration list containing
  `plant_monthly_allocation`.
}
\description{
Clean and harmonize monthly allocation inputs.
}
\details{
Reads management configuration with monthly allocation vectors and
returns a configuration containing `plant_monthly_allocation`.

Depending on the user input:
- if plant, grain and grass allocations are all provided, plant
  allocation is recalculated as grain plus grass
- if only one allocation is provided, it is treated as plant allocation
- if two allocations are provided, they must correspond to grain and
  grass; otherwise an error is returned
}
