% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomposition.R
\name{.decay}
\alias{.decay}
\title{.decay}
\usage{
.decay(CO_t, k, tempCoefficient)
}
\arguments{
\item{CO_t}{Carbon amount in the pool at time t.}

\item{k}{First-order decomposition rate constant.}

\item{tempCoefficient}{Temperature response coefficient.}
}
\value{
Carbon change due to decomposition during the timestep.
}
\description{
Calculate soil carbon loss by first-order decomposition.
}
