% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Humification_coefficient.R
\name{.hum_coef}
\alias{.hum_coef}
\title{hum_coef
humification coefficient}
\usage{
.hum_coef(f_clay)
}
\arguments{
\item{f_clay}{fraction of clay in soil}
}
\description{
hum_coef
humification coefficient
}
