% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{.pool_cn}
\alias{.pool_cn}
\title{Initial pool distribution parametrization}
\usage{
.pool_cn(cn, f_hum, f_rom, ini_Cin, soil_surf = c("top", "sub"))
}
\arguments{
\item{cn}{soil CN}

\item{f_hum_top}{initial hum fraction top layer}

\item{f_rom_top}{initial rom fraction top layer}

\item{ini_Cin_top}{initial C inputs topsoil}
}
\description{
This function helps to modify the parametrization of the initial pool distribution
according thesoil  C/N ratio.
When the C/N ratio is above the threshold of 10.8, the initial content of C in ROM is adjusted upwards, so that the relative turnover
rate is adjusted to the level determined by the function. The use of this procedure has a significant
influence for national simulations, as there is a significant proportion of coarse sandy soils in
Denmark with a high C/N ratio. If such a function is not used, the simulation of Danish sandy soils
will exhibit clear declines in SOC, in contrast to the general build-up of SOC on these soils reported
by Heidmann et al. (2001).
}
\examples{
pool_cn(cn=12,HUM_frac = 0.33, C_0=75)
}
