% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{.read_Cinputs}
\alias{.read_Cinputs}
\title{.read_Cinputs}
\usage{
.read_Cinputs(df_Cin, n)
}
\arguments{
\item{df_Cin}{dataframe with the following cols: Cin_top (residues topsoil), Cin_sub (residues subsoil), Cin_man (Manure)}

\item{n}{number of simulated years}
}
\description{
helper function to read a dataframe and return Cin_top, Cin_sub, Cin_man
}
