% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomposition.R
\name{.temp_coef}
\alias{.temp_coef}
\title{.temp_coef}
\usage{
.temp_coef(T_zt)
}
\arguments{
\item{T_zt}{Soil temperature at depth z and time t.}
}
\value{
Temperature response coefficient for carbon decomposition.
}
\description{
Calculate the temperature response coefficient for carbon decomposition.
}
