% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOM.R
\name{FOM_top_calculations}
\alias{FOM_top_calculations}
\title{FOM_top_calculations}
\usage{
FOM_top_calculations(
  FOM_top_t,
  month,
  t_avg = t_avg,
  amplitude = amplitude,
  s_config
)
}
\arguments{
\item{FOM_top_t}{Numeric. Carbon content in FOM top pool (Mg C ha-1).}

\item{month}{Integer. Month index (1–12).}

\item{t_avg}{Numeric. Monthly mean air temperature (°C).}

\item{amplitude}{Numeric. Annual historical temperature amplitude
used in the soil temperature function.}

\item{s_config}{List. Soil configuration parameters.}
}
\value{
A list with updated pool values and fluxes for FOM topsoil layer.
}
\description{
Calculate FOM decomposition and carbon fluxes in the topsoil layer.
}
\examples{
FOM_top_calculations(
  FOM_top_t = 10,
  month = 7,
  t_avg = 15,
  amplitude = 10,
  s_config = soil_config()
)

}
