% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print a greeting message}
\usage{
hello(who)
}
\value{
NULL (prints the greeting message)
}
\description{
This function prints a greeting message "Hello, world!" to the console.
}
\examples{
hello()
}
