% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{initialize_soil_pools}
\alias{initialize_soil_pools}
\title{initialize_soil_pools}
\usage{
initialize_soil_pools(cn, soil_config)
}
\arguments{
\item{cn}{Soil carbon:nitrogen ratio.}

\item{soil_config}{Soil configuration list.}
}
\value{
A list containing initialized topsoil and subsoil pools.
}
\description{
Initialize topsoil and subsoil carbon pools.
}
\examples{
s_config <- soil_config()
initialize_soil_pools(cn=15, soil_config = s_config)
}
