% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{management_config}
\alias{management_config}
\title{management_config}
\usage{
management_config(
  management_filepath = NULL,
  plant_monthly_allocation = NULL,
  grain_monthly_allocation = NULL,
  grass_monthly_allocation = NULL,
  manure_monthly_allocation = NULL,
  f_man_humification = 0.12
)
}
\arguments{
\item{management_filepath}{Either a filepath to a management template or
a data.frame containing management allocation variables.}

\item{plant_monthly_allocation}{Monthly distribution of plant carbon inputs.}

\item{grain_monthly_allocation}{Monthly distribution of grain carbon inputs.}

\item{grass_monthly_allocation}{Monthly distribution of grass carbon inputs.}

\item{manure_monthly_allocation}{Monthly distribution of manure carbon inputs.}

\item{f_man_humification}{Fraction of manure already humidified.}
}
\value{
A list containing management allocation settings.
}
\description{
Prepares management configuration for monthly carbon input allocation.
}
\details{
This function can be used in two ways:
1. from a management template file or data.frame, such as one exported with
   `export_management_template()`
2. by directly providing fixed monthly allocation vectors of length 12.

In the first approach, the user can specify monthly allocation fractions
directly in the input file. When no crop rotation is considered,
`plant_monthly_allocation` should be used. When crop rotation is considered,
grain and grass allocation fractions can be specified separately.

In the second approach, the user can directly provide monthly allocation
vectors of length 12.
}
\examples{
management_config(
f_man_humification=0.192,
plant_monthly_allocation = c(0,0,0,.08,.12,.16,.64,0,0,0,0,0),
 manure_monthly_allocation = c(0,0,1,0,0,0,0,0,0,0,0,0)
 )
}
