% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature.R
\name{prepare_temperature}
\alias{prepare_temperature}
\title{prepare_temperature}
\usage{
prepare_temperature(
  filepath,
  Tavg_col = "Tavg",
  Trange_col = NULL,
  default_Trange = 15
)
}
\arguments{
\item{filepath}{}

\item{Tavg_col}{}

\item{Trange_col}{}

\item{default_Trange}{}
}
\description{
prepare_temperature
}
