% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario}
\alias{scenario}
\title{rCTOOL scenario example dataset}
\format{
A data frame.
}
\source{
Internal example dataset distributed with rCTOOL.
}
\usage{
scenario
}
\description{
Example dataset used to demonstrate scenario simulations in rCTOOL.
}
\keyword{datasets}
