% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_temperature}
\alias{scenario_temperature}
\title{rCTOOL scenario temperature dataset}
\format{
A data frame.
}
\source{
Internal example dataset distributed with rCTOOL.
}
\usage{
scenario_temperature
}
\description{
Monthly temperature dataset used in rCTOOL scenario simulations.
}
\keyword{datasets}
