% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperature.R
\name{set_monthly_temperature_data}
\alias{set_monthly_temperature_data}
\title{set_monthly_temperature_data}
\usage{
set_monthly_temperature_data(file = NULL, Tavg = NULL)
}
\arguments{
\item{file}{Either a filepath to a csv file or a data.frame containing
monthly temperature data.}

\item{Tavg}{Numeric vector of average monthly temperature values.}
}
\value{
A data.frame containing monthly temperature input data.
  The output always contains `Tavg`.
}
\description{
Prepare monthly temperature input data for rCTOOL.
}
\details{
Monthly temperature data can be provided in two ways:
1. directly through the argument `file`, as a filepath or data.frame
   containing monthly temperature data
2. directly through the argument `Tavg`, as a numeric vector of average
   monthly temperature values
}
\examples{
df_temp <- data.frame(Tavg = c(5, 6, 7, 8, 9, 15, 14, 11, 10, 9, 5, 1))
set_monthly_temperature_data(file = df_temp)

set_monthly_temperature_data(
  Tavg = c(5, 6, 7, 8, 9, 15, 14, 11, 10, 9, 5, 1)
)
}
