% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomposition.R
\name{soil_pool_decomposition}
\alias{soil_pool_decomposition}
\title{soil_pool_decomposition}
\usage{
soil_pool_decomposition(
  soil_pool,
  k,
  soil_depth,
  month,
  t_avg,
  amplitude,
  s_config
)
}
\arguments{
\item{soil_pool}{Carbon amount in the soil pool.}

\item{k}{First-order decomposition rate constant.}

\item{soil_depth}{Soil layer thickness in cm. In the default C-TOOL
configuration, the model operates with two fixed layers: topsoil
(25 cm) and subsoil (100 cm).}

\item{month}{Month index from 1 to 12.}

\item{t_avg}{Monthly mean air temperature.}

\item{amplitude}{Historical annual temperature amplitude previously
calculate from the monthly temperature series.}

\item{s_config}{Soil parameter configuration.}
}
\value{
Carbon change in the pool due to temperature-modified
first-order decomposition.
}
\description{
Calculate temperature-modified decomposition of a soil carbon pool.
}
