% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physical_boundaries.R
\name{check_balance}
\alias{check_balance}
\title{check_balance}
\usage{
check_balance(ctool_output, cin_config, s_config)
}
\arguments{
\item{ctool_output}{output from turnover model}

\item{cin_config}{carbon input config}

\item{s_config}{soil config}
}
\description{
calculates mass balance for the simulation
Includes initial c soil, sum of inputs, SOC stocks and CO2 emissions
needs to be 0 or really close to 0
used in run_ctool (ctool.R)
}
