% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{define_Cinputs}
\alias{define_Cinputs}
\title{define_Cinputs}
\usage{
define_Cinputs(
  df_Cin = NULL,
  Cin_top = NULL,
  Cin_sub = NULL,
  Cin_man = NULL,
  time_config
)
}
\arguments{
\item{df_Cin}{dataframe with the following cols: Cin_top (residues topsoil), Cin_sub (residues subsoil), Cin_man (Manure)}

\item{Cin_top}{C input from residues on topsoil}

\item{Cin_sub}{C input from residues on subsoil}

\item{Cin_man}{C input from manure}

\item{time_config}{config of timeperiod}
}
\description{
explicits C inputs from plants and manure
}
