% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Physical_boundaries.R
\name{define_physical_boundaries}
\alias{define_physical_boundaries}
\title{define_physical_boundaries}
\usage{
define_physical_boundaries(value, min_limit, max_limit)
}
\arguments{
\item{value}{parameter value}

\item{min_limit}{min limit (usually 0)}

\item{max_limit}{max limit}
}
\description{
establishes minimum and physical boundaries for parameters
}
\examples{
define_physical_boundaries(0.3, 0, 1)
}
