% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{define_timeperiod}
\alias{define_timeperiod}
\title{define_timeperiod}
\usage{
define_timeperiod(yr_start = 2006, yr_end = 2010)
}
\arguments{
\item{yr_start}{initial simulation year}

\item{yr_end}{end simulation year}
}
\description{
creates a
}
