% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_soil_pools.R
\name{.clean_monthly_allocations}
\alias{.clean_monthly_allocations}
\title{.clean_monthly_allocations}
\usage{
.clean_monthly_allocations(m_config)
}
\arguments{
\item{m_config}{management config

 @description
 Reads management config with monthly allocations (plant, grain/grass)
 Depending on the user specifications, adapts the management config:
 (i) if user specifies plant, grain and grass allocations, recalculate plant allocation (sum of grain and grass)
 (ii) if only one allocation is given, see whether it is plant (no crop rotation), provides user with a confirmation message
 (iii) if two allocation are given, confirm these match crop rotation (grass/grain), otherwise stop}
}
\value{
Returns only plant monthly allocation irrespective of user input; this will be read with fucntions bellow
}
\description{
.clean_monthly_allocations
}
