% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomposition.R
\name{.decay}
\alias{.decay}
\title{decay
Decay of carbon function'
The decay of carbon in each pool is described by first-order reaction kinetics
where we have an specific decay rate coefficient in each pool that affects the
actual C content in it (MgC/ha).The turnover is simulated in monthly time steps,
where th decay rate is modify by the temp_coef() function.}
\usage{
.decay(CO_t, k, tempCoefficient)
}
\arguments{
\item{CO_t}{description}

\item{k}{description}

\item{tempCoefficient}{description}
}
\value{
T
}
\description{
decay
Decay of carbon function'
The decay of carbon in each pool is described by first-order reaction kinetics
where we have an specific decay rate coefficient in each pool that affects the
actual C content in it (MgC/ha).The turnover is simulated in monthly time steps,
where th decay rate is modify by the temp_coef() function.
}
