% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOM.R
\name{FOM_sub_calculations}
\alias{FOM_sub_calculations}
\title{FOM_sub_calculations}
\usage{
FOM_sub_calculations(
  FOM_sub_t,
  month,
  t_avg = t_avg,
  t_range = t_range,
  s_config
)
}
\arguments{
\item{FOM_sub_t}{}

\item{month}{}

\item{t_avg}{}

\item{t_range}{}

\item{s_config}{}
}
\description{
FOM_sub_calculations
}
