% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUM.R
\name{HUM_sub_calculations}
\alias{HUM_sub_calculations}
\title{HUM_sub_calculations}
\usage{
HUM_sub_calculations(
  HUM_sub_t,
  month,
  t_avg = t_avg,
  t_range = t_range,
  s_config
)
}
\arguments{
\item{s_config}{}
}
\description{
HUM_sub_calculations
}
