% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUM.R
\name{HUM_top_calculations}
\alias{HUM_top_calculations}
\title{HUM_top_calculations}
\usage{
HUM_top_calculations(
  HUM_top_t,
  month,
  t_avg = t_avg,
  t_range = t_range,
  s_config
)
}
\arguments{
\item{s_config}{}
}
\description{
HUM_top_calculations
}
