% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{management_config}
\alias{management_config}
\title{management_config}
\usage{
management_config(
  f_man_humification = 0.192,
  plant_monthly_allocation = c(0, 0, 0, 0.08, 0.12, 0.16, 0.64, 0, 0, 0, 0, 0),
  manure_monthly_allocation = c(0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0)
)
}
\arguments{
\item{f_man_humification}{fraction of manure already humidified}

\item{plant_monthly_allocation}{monthly distribution of plant C inputs}

\item{manure_monthly_allocation}{monthly distribution of manure C input}
}
\description{
management_config
}
