% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCTOOL-package.R
\docType{package}
\name{rCTOOL-package}
\alias{rCTOOL}
\alias{rCTOOL-package}
\title{rCTOOL: An open and comprehensive R package for C turnover modelling}
\description{
rCTOOL is an open-source R package, that encapsulates the capabilities of C-TOOL while addressing its implementation limitations. The aim consists of providing a user-friendly interface, facilitates running multiple scenarios, and offers comprehensive documentation and licensing information. This package streamlines the use of C-TOOL, making it more accessible and effective for potential users.
}
\keyword{internal}
