% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROM.R
\name{ROM_sub_calculations}
\alias{ROM_sub_calculations}
\title{ROM_sub_calculations}
\usage{
ROM_sub_calculations(
  ROM_sub_t,
  month,
  t_avg = t_avg,
  t_range = t_range,
  s_config
)
}
\arguments{
\item{ROM_sub_t}{}

\item{month}{}

\item{t_avg}{}

\item{t_range}{}

\item{s_config}{}
}
\description{
ROM_sub_calculations
}
