% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROM.R
\name{ROM_top_calculations}
\alias{ROM_top_calculations}
\title{ROM_top_calculations}
\usage{
ROM_top_calculations(
  ROM_top_t,
  month,
  t_avg = t_avg,
  t_range = t_range,
  s_config
)
}
\arguments{
\item{ROM_top_t}{}

\item{month}{}

\item{t_avg}{}

\item{t_range}{}

\item{s_config}{}
}
\description{
ROM_top_calculations
}
