% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameter_config.R
\name{soil_config}
\alias{soil_config}
\title{soil_config}
\usage{
soil_config(
  Csoil_init = 70.4,
  f_hum_top = 0.48,
  f_rom_top = 0.49,
  f_hum_sub = 0.312,
  f_rom_sub = 0.6847,
  Cproptop = 0.47,
  clay_top = 0.1,
  clay_sub = 0.15,
  phi = 0.035,
  f_co2 = 0.628,
  f_romi = 0.012,
  k_fom = 0.12,
  k_hum = 0.0028,
  k_rom = 3.85e-05,
  ftr = 0.0025,
  cn = 7.166667
)
}
\arguments{
\item{Csoil_init}{initial C stock at depth 1m (t/ha)}

\item{f_hum_top}{initial hum fraction top layer}

\item{f_rom_top}{initial rom fraction top layer}

\item{f_hum_sub}{initial hum fraction bottom layer}

\item{f_rom_sub}{initial rom fraction bottom layer}

\item{Cproptop}{Proportion of the total C allocated in topsoil}

\item{clay_top}{clay fraction top soil}

\item{clay_sub}{clay fraction subsoil}

\item{phi}{Diffusion index}

\item{f_co2}{respiration fraction}

\item{f_romi}{romification fraction}

\item{k_fom}{fom decomposition rate}

\item{k_hum}{hum decomposition rate}

\item{k_rom}{rom decomposition rate}

\item{ftr}{transport rate}

\item{cn}{soil Carbon:Nitrogen ratio}

\item{ini_Cin_top}{initial C inputs topsoil}

\item{ini_Cin_sub}{initial C inputs subsoil}
}
\description{
soil configuration params
}
