% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomposition.R
\name{soil_pool_decomposition}
\alias{soil_pool_decomposition}
\title{soil_pool_decomposition}
\usage{
soil_pool_decomposition(
  soil_pool = c("FOM_top", "FOM_sub", "HUM_top", "HUM_sub", "ROM_top", "ROM_sub"),
  k,
  soil_depth = c(25, 100),
  month,
  t_avg,
  t_range,
  s_config
)
}
\arguments{
\item{soil_pool}{}

\item{k}{}

\item{soil_depth}{}

\item{month}{}

\item{t_avg}{}

\item{t_range}{}

\item{s_config}{}
}
\description{
soil_pool_decomposition
}
