% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctool.R, R/turnover_beta.R
\name{turnover}
\alias{turnover}
\title{turnover-rCTOOL}
\usage{
turnover(i)

turnover(i)
}
\arguments{
\item{timestep}{}

\item{timeperiod}{}

\item{cin_config}{}

\item{m_config}{}

\item{t_config}{}

\item{s_config}{}

\item{out}{}
}
\value{
The arguments are:
}
\description{
turnover-rCTOOL

Turnover core function
}
