% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_soil_pools.R
\name{update_monthly_FOM_sub}
\alias{update_monthly_FOM_sub}
\title{updated_monthly_FOM_sub}
\usage{
update_monthly_FOM_sub(
  FOM_sub_t1,
  FOM_transport,
  C_in_plant_sub,
  month,
  m_config
)
}
\arguments{
\item{FOM_sub_t1}{}

\item{FOM_transport}{}

\item{C_in_plant_sub}{}

\item{month}{}

\item{m_config}{}
}
\description{
updated_monthly_FOM_sub
}
