% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_soil_pools.R
\name{update_monthly_FOM_top}
\alias{update_monthly_FOM_top}
\title{update_FOM_top}
\usage{
update_monthly_FOM_top(FOM_top_t1, Cin_plant_top, Cin_manure, month, m_config)
}
\arguments{
\item{m_config}{management configuration list}
}
\description{
update_FOM_top
}
