% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_soil_pools.R
\name{update_monthly_HUM_sub}
\alias{update_monthly_HUM_sub}
\title{update_monthly_HUM_sub}
\usage{
update_monthly_HUM_sub(HUM_sub_t1, HUM_transport, FOM_humified_sub)
}
\arguments{
\item{FOM_humified_sub}{}
}
\description{
update_monthly_HUM_sub
}
