% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_soil_pools.R
\name{update_monthly_HUM_top}
\alias{update_monthly_HUM_top}
\title{update_monthly_HUM_top}
\usage{
update_monthly_HUM_top(HUM_top_t1, C_in_man, FOM_humified_top, month, m_config)
}
\arguments{
\item{m_config}{}
}
\description{
update_monthly_HUM_top
}
