% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_soil_pools.R
\name{update_monthly_ROM_sub}
\alias{update_monthly_ROM_sub}
\title{update_monthly_ROM_sub}
\usage{
update_monthly_ROM_sub(ROM_sub_t1, HUM_romified_sub, ROM_transport)
}
\arguments{
\item{ROM_transport}{}
}
\description{
update_monthly_ROM_sub
}
