% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Update_soil_pools.R
\name{update_monthly_ROM_top}
\alias{update_monthly_ROM_top}
\title{update_monthly_ROM_top}
\usage{
update_monthly_ROM_top(ROM_top_t1, HUM_romified_top)
}
\arguments{
\item{ROM_top_t1}{}

\item{HUM_romified_top}{}
}
\description{
update_monthly_ROM_top
}
